added test validating implementation of spherical harmonics evaluation. This implementation is thus validated

+++ doxygen_script.sh
@@ -1,7 +1,11 @@
-cd doc/
-git rm -rf *
-git commit -m "purge before updates"
-doxygen ../Doxyfile
-git add *
-git commit -m "updates to documentation"
-git push origin gh-pages
+
+if cd doc/; then
+	git rm -rf *
+	git commit -m "purge before updates"
+	doxygen ../Doxyfile
+	git add *
+	git commit -m "updates to documentation"
+	git push origin gh-pages
+else
+	echo "Command failed"
+fi
